requirements.txt
pandas
numpy
requests
flask
discord.py
scikit-learn
statsmodels
scipy
psutil
python-dotenv
